requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=1.0=mkl
ca-certificates=2020.1.1=0
certifi=2019.11.28=py37_0
cloudpickle=1.3.0=py_0
cycler=0.10.0=py37_0
cytoolz=0.10.1=py37h1de35cc_0
dask-core=2.11.0=py_0
dbus=1.13.12=h90a0687_0
decorator=4.4.1=py_0
expat=2.2.6=h0a44026_0
freetype=2.9.1=hb4e5f40_0
gettext=0.19.8.1=h15daf44_3
glib=2.63.1=hd977a24_0
icu=58.2=h4b95b61_1
imageio=2.6.1=py37_0
intel-openmp=2019.4=233
jpeg=9b=he5867d9_2
kiwisolver=1.1.0=py37h0a44026_0
libcxx=4.0.1=hcfea43d_1
libcxxabi=4.0.1=hcfea43d_1
libedit=3.1.20181209=hb402a30_0
libffi=3.2.1=h475c297_4
libgfortran=3.0.1=h93005f0_2
libiconv=1.15=hdd342a3_7
libpng=1.6.37=ha441bb4_0
libtiff=4.1.0=hcb84e12_0
matplotlib=3.1.3=py37_0
matplotlib-base=3.1.3=py37h9aa3819_0
mkl=2019.4=233
mkl-service=2.3.0=py37hfbe908c_0
mkl_fft=1.0.15=py37h5e564d8_0
mkl_random=1.1.0=py37ha771720_0
ncurses=6.2=h0a44026_0
networkx=2.4=py_0
numpy=1.18.1=py37h7241aed_0
numpy-base=1.18.1=py37h6575580_1
olefile=0.46=py37_0
openssl=1.1.1d=h1de35cc_4
pcre=8.43=h0a44026_0
pillow=7.0.0=py37h4655f20_0
pip=20.0.2=py37_1
pyparsing=2.4.6=py_0
pyqt=5.9.2=py37h655552a_2
python=3.7.6=h359304d_2
python-dateutil=2.8.1=py_0
pywavelets=1.1.1=py37h1de35cc_0
qt=5.9.7=h468cd18_1
readline=7.0=h1de35cc_5
scikit-image=0.16.2=py37h6c726b0_0
scipy=1.4.1=py37h9fa6033_0
setuptools=45.2.0=py37_0
sip=4.19.8=py37h0a44026_0
six=1.14.0=py37_0
sqlite=3.31.1=ha441bb4_0
tk=8.6.8=ha441bb4_0
toolz=0.10.0=py_0
tornado=6.0.3=py37h1de35cc_3
wheel=0.34.2=py37_0
xz=5.2.4=h1de35cc_4
zlib=1.2.11=h1de35cc_3
zstd=1.3.7=h5bba6e5_0
